+++ CHANGELOG.md
@@ -5,12 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Added
+
+ - Allows to pass a list of xmlid to import
+
 ## [3.6.1] - 2024-09-30
 
 ### Added
 
- - Allows to add several connections to different Odoo database
- - Allows to add sql connections to PostgreSQL and MSSQL databases and use them in script files
+ - Allows to add several connections to different Odoo databases
+ - Allows to add sql connections to PostgresSQL and MSSQL databases and use them in script files
 
 ## [3.6.0] - 2024-09-25